Revert "source-postgres: enable ctid+cdc implementation (#28044)"
This reverts commit 886ab6b09b95ec64ac24d5a0688ab08c931bd938.
