+++ airbyte-integrations/connectors/source-postgres/src/main/java/io/airbyte/integrations/source/postgres/ctid/CtidFeatureFlags.java
@@ -10,11 +10,16 @@ import com.fasterxml.jackson.databind.JsonNode;
 // One for each type: CDC and standard cursor based
 public class CtidFeatureFlags {
 
+  public static final String CDC_VIA_CTID = "cdc_via_ctid";
   public static final String CURSOR_VIA_CTID = "cursor_via_ctid";
   private final JsonNode sourceConfig;
 
   public CtidFeatureFlags(final JsonNode sourceConfig) {
     this.sourceConfig = sourceConfig;
+  }
+
+  public boolean isCdcSyncEnabled() {
+    return getFlagValue(CDC_VIA_CTID);
   }
 
   public boolean isCursorSyncEnabled() {
